feat: fetch online data while scanning data
